+++ src/infrastructure/dbcontroller.py
@@ -13,7 +13,7 @@ from stat import S_IXGRP, S_IXOTH, S_IXUSR
 
 import mysql.connector
 
-from domain.const import MODULE_LOG_NAME, MYSQL_CONFIG
+from infrastructure.const import MODULE_LOG_NAME, MYSQL_CONFIG
 from domain.dbcontroller import DatabaseController
