+++ src/module/getdata.py
@@ -15,7 +15,7 @@ import bs4
 import numpy as np
 import pandas as pd
 
-from domain import const
+from infrastructure import const
 from domain.model.info import (ChokuzenPlayerInfo, ProgramCommonInfo,
                                ProgramPlayerInfo, ResultCommonInfo,
                                ResultPlayerInfo, Tansho, ThreeRenfuku,
